requirements.txt
pendulum
colorama
fpdf2